Jwt implementado na rota user

+++ src/services/UserService.ts
@@ -1,18 +1,21 @@
 import UserModel from '../models/UserModel';
-import { IUser } from '../types/IUser';
+import { INewUser, IUser } from '../types/IUser';
 import { ServiceResponse } from '../types/ServiceResponseTypes';
+import JwtUtils from '../utils/JwtUtils';
 import * as bcrypt from 'bcryptjs';
 
 export default class UserService {
   private userModel: UserModel;
+  private jwtUtils;
   private emailInUse = "Email address is already in use by another user."
   private userNotFound = "User not found"
 
   constructor() {
     this.userModel = new UserModel();
+    this.jwtUtils = new JwtUtils();
   }
 
-  async create(user: IUser): Promise<ServiceResponse< Partial<IUser> | null>> {
+  async create(user: INewUser): Promise<ServiceResponse< Partial<IUser> | null>> {
     try {
       const encryptedPassword = await bcrypt.hash(user.password, 10);
       const newUser = { ...user, password: encryptedPassword }
@@ -24,13 +27,26 @@ export default class UserService {
     }
   }
 
+  async login(email: string, password: string): Promise<ServiceResponse<{ token: string }>> {
+    const retrievedUser = await this.userModel.findByEmail(email);
+    
+    if (!retrievedUser || await !bcrypt.compare(password, retrievedUser.password!)) {
+      return { status: 'INVALID_DATA', data: { message: 'Invalid email or password' } };
+    }
+    const token = this.jwtUtils.sign({ id: retrievedUser.id!, email: retrievedUser.email!, role: retrievedUser.role! });
+    return { status: 'SUCCESSFUL', data: { token } };
+  }
+
   async findById(id: string): Promise<ServiceResponse<Partial<IUser>>> {
     const user = await this.userModel.findById(id);
     if (!user) return { status: 'NOT_FOUND', data: { message: this.userNotFound } };
     return { status: 'SUCCESSFUL', data: user };
   }
 
-  async update(id: string, user: IUser): Promise<ServiceResponse<Partial<IUser> | null>> {
+  async update(id: string, user: INewUser, token: string): Promise<ServiceResponse<Partial<IUser> | null>> {
+    if (!this.jwtUtils.isAdmin(token) && id !== this.jwtUtils.decode(token).id) {
+      return { status: 'UNAUTHORIZED', data: { message: 'You are not authorized to update this user' } };
+    }
     try {
       const encryptedPassword = await bcrypt.hash(user.password, 10);
       const newUser = { ...user, password: encryptedPassword }
@@ -44,7 +60,10 @@ export default class UserService {
     }
   }
 
-  async delete(id: string): Promise<ServiceResponse<IUser>> {
+  async delete(id: string, token: string): Promise<ServiceResponse<IUser>> {
+    if (!this.jwtUtils.isAdmin(token) && id !== this.jwtUtils.decode(token).id) {
+      return { status: 'UNAUTHORIZED', data: { message: 'You are not authorized to delete this user' } };
+    }
     const deletedUser = await this.userModel.delete(id);
     if (!deletedUser) return { status: 'NOT_FOUND', data: { message: this.userNotFound } };
     return { status: 'SUCCESSFUL', data: deletedUser };
